reverting next server side instrumentation refactor
