--- common/api-review/crashlytics-next.api.md
@@ -1,31 +1,0 @@
-## API Report File for "@firebase/crashlytics-next"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { FirebaseApp } from '@firebase/app';
-import { Instrumentation } from 'next';
-
-// @public
-export interface Crashlytics {
-    app: FirebaseApp;
-}
-
-// @public
-export interface CrashlyticsOptions {
-    appVersion?: string;
-    endpointUrl?: string;
-    region?: string;
-    tracingUrl?: string;
-}
-
-export { Instrumentation }
-
-// @public
-export function nextOnRequestError(crashlyticsOptions?: CrashlyticsOptions): Instrumentation.onRequestError;
-
-
-// (No @packageDocumentation comment for this package)
-
-```
