+++ docs-devsite/_toc.yaml
@@ -456,13 +456,6 @@ toc:
     path: /docs/reference/js/crashlytics_angular.crashlyticsoptions.md
   - title: FirebaseErrorHandler
     path: /docs/reference/js/crashlytics_angular.firebaseerrorhandler.md
-- title: crashlytics/next
-  path: /docs/reference/js/crashlytics_next.md
-  section:
-  - title: Crashlytics
-    path: /docs/reference/js/crashlytics_next.crashlytics.md
-  - title: CrashlyticsOptions
-    path: /docs/reference/js/crashlytics_next.crashlyticsoptions.md
 - title: crashlytics/next-navigation
   path: /docs/reference/js/crashlytics_next-navigation.md
   section:
