--- docs-devsite/crashlytics_next.crashlytics.md
@@ -1,37 +1,0 @@
-Project: /docs/reference/js/_project.yaml
-Book: /docs/reference/_book.yaml
-page_type: reference
-
-{% comment %}
-DO NOT EDIT THIS FILE!
-This is generated by the JS SDK team, and any local changes will be
-overwritten. Changes should be made in the source code at
-https://github.com/firebase/firebase-js-sdk
-{% endcomment %}
-
-# Crashlytics interface
-An instance of the Firebase Crashlytics SDK.
-
-Do not create this instance directly. Instead, use [getCrashlytics()](./crashlytics_.md#getcrashlytics_a9d22a1)<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-export interface Crashlytics 
-```
-
-## Properties
-
-|  Property | Type | Description |
-|  --- | --- | --- |
-|  [app](./crashlytics_next.crashlytics.md#crashlyticsapp) | [FirebaseApp](./app.firebaseapp.md#firebaseapp_interface) | The [FirebaseApp](./app.firebaseapp.md#firebaseapp_interface) this [Crashlytics](./crashlytics_.crashlytics.md#crashlytics_interface) instance is associated with. |
-
-## Crashlytics.app
-
-The [FirebaseApp](./app.firebaseapp.md#firebaseapp_interface) this [Crashlytics](./crashlytics_.crashlytics.md#crashlytics_interface) instance is associated with.
-
-<b>Signature:</b>
-
-```typescript
-app: FirebaseApp;
-```
