--- docs-devsite/crashlytics_next.crashlyticsoptions.md
@@ -1,72 +1,0 @@
-Project: /docs/reference/js/_project.yaml
-Book: /docs/reference/_book.yaml
-page_type: reference
-
-{% comment %}
-DO NOT EDIT THIS FILE!
-This is generated by the JS SDK team, and any local changes will be
-overwritten. Changes should be made in the source code at
-https://github.com/firebase/firebase-js-sdk
-{% endcomment %}
-
-# CrashlyticsOptions interface
-Options for initializing the Crashlytics service using [getCrashlytics()](./crashlytics_.md#getcrashlytics_a9d22a1)<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-export interface CrashlyticsOptions 
-```
-
-## Properties
-
-|  Property | Type | Description |
-|  --- | --- | --- |
-|  [appVersion](./crashlytics_next.crashlyticsoptions.md#crashlyticsoptionsappversion) | string | The version of the application. This should be a unique string that identifies the snapshot of code to be deployed, such as "1.0.2". If not specified, other default locations will be checked for an identifier. Setting a value here takes precedence over any other values. |
-|  [endpointUrl](./crashlytics_next.crashlyticsoptions.md#crashlyticsoptionsendpointurl) | string | The URL for the endpoint to which Crashlytics data should be sent, in the OpenTelemetry format. By default, data will be sent to Firebase. |
-|  [region](./crashlytics_next.crashlyticsoptions.md#crashlyticsoptionsregion) | string | The Google Cloud region where the Crashlytics data should be sent.<!-- -->By default, data will be sent to the "global" region.<!-- -->Refer to https://cloud.google.com/logging/docs/regions for the list of available regions. |
-|  [tracingUrl](./crashlytics_next.crashlyticsoptions.md#crashlyticsoptionstracingurl) | string | The URL for the endpoint to which Crashlytics traces should be sent, in the OpenTelemetry format. By default, data will be sent to Firebase. |
-
-## CrashlyticsOptions.appVersion
-
-The version of the application. This should be a unique string that identifies the snapshot of code to be deployed, such as "1.0.2". If not specified, other default locations will be checked for an identifier. Setting a value here takes precedence over any other values.
-
-<b>Signature:</b>
-
-```typescript
-appVersion?: string;
-```
-
-## CrashlyticsOptions.endpointUrl
-
-The URL for the endpoint to which Crashlytics data should be sent, in the OpenTelemetry format. By default, data will be sent to Firebase.
-
-<b>Signature:</b>
-
-```typescript
-endpointUrl?: string;
-```
-
-## CrashlyticsOptions.region
-
-The Google Cloud region where the Crashlytics data should be sent.
-
-By default, data will be sent to the "global" region.
-
-Refer to https://cloud.google.com/logging/docs/regions for the list of available regions.
-
-<b>Signature:</b>
-
-```typescript
-region?: string;
-```
-
-## CrashlyticsOptions.tracingUrl
-
-The URL for the endpoint to which Crashlytics traces should be sent, in the OpenTelemetry format. By default, data will be sent to Firebase.
-
-<b>Signature:</b>
-
-```typescript
-tracingUrl?: string;
-```
