--- docs-devsite/crashlytics_next.instrumentation.md
@@ -1,31 +1,0 @@
-Project: /docs/reference/js/_project.yaml
-Book: /docs/reference/_book.yaml
-page_type: reference
-
-{% comment %}
-DO NOT EDIT THIS FILE!
-This is generated by the JS SDK team, and any local changes will be
-overwritten. Changes should be made in the source code at
-https://github.com/firebase/firebase-js-sdk
-{% endcomment %}
-
-# Instrumentation namespace
-<b>Signature:</b>
-
-```typescript
-export declare namespace Instrumentation 
-```
-
-## Type Aliases
-
-|  Type Alias | Description |
-|  --- | --- |
-|  [onRequestError](./crashlytics_next.instrumentation.md#instrumentationonrequesterror) |  |
-
-## Instrumentation.onRequestError
-
-<b>Signature:</b>
-
-```typescript
-type onRequestError = InstrumentationOnRequestError;
-```
