--- docs-devsite/crashlytics_next.md
@@ -1,66 +1,0 @@
-Project: /docs/reference/js/_project.yaml
-Book: /docs/reference/_book.yaml
-page_type: reference
-
-{% comment %}
-DO NOT EDIT THIS FILE!
-This is generated by the JS SDK team, and any local changes will be
-overwritten. Changes should be made in the source code at
-https://github.com/firebase/firebase-js-sdk
-{% endcomment %}
-
-# @firebase/crashlytics/next
-
-## Functions
-
-|  Function | Description |
-|  --- | --- |
-|  [nextOnRequestError(crashlyticsOptions)](./crashlytics_next.md#nextonrequesterror_3caf5de) | Automatically report uncaught errors from server routes to Firebase Crashlytics. |
-
-## Interfaces
-
-|  Interface | Description |
-|  --- | --- |
-|  [Crashlytics](./crashlytics_next.crashlytics.md#crashlytics_interface) | An instance of the Firebase Crashlytics SDK.<!-- -->Do not create this instance directly. Instead, use [getCrashlytics()](./crashlytics_.md#getcrashlytics_a9d22a1)<!-- -->. |
-|  [CrashlyticsOptions](./crashlytics_next.crashlyticsoptions.md#crashlyticsoptions_interface) | Options for initializing the Crashlytics service using [getCrashlytics()](./crashlytics_.md#getcrashlytics_a9d22a1)<!-- -->. |
-
-## Namespaces
-
-|  Namespace | Description |
-|  --- | --- |
-|  [Instrumentation](./crashlytics_next.instrumentation.md#instrumentation_namespace) |  |
-
-## function(crashlyticsOptions, ...)
-
-### nextOnRequestError(crashlyticsOptions) {:#nextonrequesterror_3caf5de}
-
-Automatically report uncaught errors from server routes to Firebase Crashlytics.
-
-<b>Signature:</b>
-
-```typescript
-export declare function nextOnRequestError(crashlyticsOptions?: CrashlyticsOptions): Instrumentation.onRequestError;
-```
-
-#### Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  crashlyticsOptions | [CrashlyticsOptions](./crashlytics_.crashlyticsoptions.md#crashlyticsoptions_interface) | [CrashlyticsOptions](./crashlytics_.crashlyticsoptions.md#crashlyticsoptions_interface) that configure the Crashlytics instance. |
-
-<b>Returns:</b>
-
-[Instrumentation.onRequestError](./crashlytics_.instrumentation.md#instrumentationonrequesterror)
-
-A request error handler for use in Next.js' instrumentation file
-
-### Example
-
-
-```javascript
-// In instrumentation.ts (https://nextjs.org/docs/app/guides/instrumentation):
-import { nextOnRequestError } from 'firebase/crashlytics'
-export const onRequestError = nextOnRequestError();
-
-```
-
