+++ packages/crashlytics/api-extractor.json
@@ -15,10 +15,6 @@
       "filePath": "<projectFolder>/dist/react-router/index.d.ts"
     },
     {
-      "modulePath": "next",
-      "filePath": "<projectFolder>/dist/next/index.d.ts"
-    },
-    {
       "modulePath": "next-navigation",
       "filePath": "<projectFolder>/dist/next-navigation/index.d.ts"
     }
