+++ packages/crashlytics/package.json
@@ -55,18 +55,6 @@
       },
       "default": "./dist/angular/index.esm.js"
     },
-    "./next": {
-      "types": "./dist/next/index-public.d.ts",
-      "node": {
-        "import": "./dist/next/index.esm.js",
-        "default": "./dist/next/index.cjs.js"
-      },
-      "browser": {
-        "require": "./dist/next/index.cjs.js",
-        "import": "./dist/next/index.esm.js"
-      },
-      "default": "./dist/next/index.esm.js"
-    },
     "./next-navigation": {
       "types": "./dist/next-navigation/index-public.d.ts",
       "node": {
@@ -98,12 +86,11 @@
     "test:browser": "karma start",
     "test:node": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' nyc --reporter lcovonly -- mocha **/*.test.ts src/**/*.test.ts --exclude src/angular/*.test.ts --config ../../config/mocharc.node.js",
     "trusted-type-check": "tsec -p tsconfig.json --noEmit",
-    "api-report": "yarn api-report:main && yarn api-report:react && yarn api-report:react-router && yarn api-report:angular && yarn api-report:next && yarn api-report:next-navigation && yarn api-report:api-json",
+    "api-report": "yarn api-report:main && yarn api-report:react && yarn api-report:react-router && yarn api-report:angular && yarn api-report:next-navigation && yarn api-report:api-json",
     "api-report:main": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics --packageRoot . --typescriptDts ./dist/index.d.ts --rollupDts ./dist/private.d.ts --untrimmedRollupDts ./dist/internal.d.ts --publicDts ./dist/index-public.d.ts",
     "api-report:react": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics-react --packageRoot . --typescriptDts ./dist/react/index.d.ts --rollupDts ./dist/react/private.d.ts --untrimmedRollupDts ./dist/react/internal.d.ts --publicDts ./dist/react/index-public.d.ts",
     "api-report:react-router": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics-react-router --packageRoot . --typescriptDts ./dist/react-router/index.d.ts --rollupDts ./dist/react-router/private.d.ts --untrimmedRollupDts ./dist/react-router/internal.d.ts --publicDts ./dist/react-router/index-public.d.ts",
     "api-report:angular": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics-angular --packageRoot . --typescriptDts ./dist/angular/index.d.ts --rollupDts ./dist/angular/private.d.ts --untrimmedRollupDts ./dist/angular/internal.d.ts --publicDts ./dist/angular/index-public.d.ts",
-    "api-report:next": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics-next --packageRoot . --typescriptDts ./dist/next/index.d.ts --rollupDts ./dist/next/private.d.ts --untrimmedRollupDts ./dist/next/internal.d.ts --publicDts ./dist/next/index-public.d.ts",
     "api-report:next-navigation": "ts-node-script ../../repo-scripts/prune-dts/extract-public-api.ts --package crashlytics-next-navigation --packageRoot . --typescriptDts ./dist/next-navigation/index.d.ts --rollupDts ./dist/next-navigation/private.d.ts --untrimmedRollupDts ./dist/next-navigation/internal.d.ts --publicDts ./dist/next-navigation/index-public.d.ts",
     "api-report:api-json": "api-extractor run --local --verbose",
     "typings:public": "node ../../scripts/build/use_typings.js ./dist/index-public.d.ts"
